refactor: refactored bookmark page
