+++ packages/widget/src/components/BookmarkLanding/BookmarkLanding.tsx
@@ -1,0 +1,32 @@
+import Alert from '@mui/material/Alert';
+import { useTranslation } from 'react-i18next';
+import { Card } from '../../components/Card';
+import { FormControl, Input, PageContainer } from './BookmarkLanding.styled';
+
+export const BookmarkLanding = () => {
+  const { t } = useTranslation();
+  return (
+    <PageContainer>
+      <Alert severity="info">{t('info.message.fundsToExchange')}</Alert>
+      <Card sx={{ marginTop: '16px' }}>
+        <FormControl
+          fullWidth
+          sx={{
+            paddingTop: '6px',
+            paddingBottom: '5px',
+          }}
+        >
+          <Input
+            sx={{ height: '96px', alignItems: 'flex-start' }}
+            size="small"
+            autoComplete="off"
+            autoCorrect="off"
+            autoCapitalize="off"
+            spellCheck="false"
+            placeholder={t('bookmark.enterAddressOrEns') as string}
+          />
+        </FormControl>
+      </Card>
+    </PageContainer>
+  );
+};
