+++ packages/widget/src/components/BookmarkLanding/index.ts
@@ -1,0 +1,1 @@
+export * from './BookmarkLanding'
