+++ packages/widget/src/pages/WalletBookmark/WalletBookmark.tsx
@@ -1,32 +1,7 @@
-import Alert from '@mui/material/Alert';
-import { useTranslation } from 'react-i18next';
-import { Card } from '../../components/Card';
-import { FormControl, Input, PageContainer } from './WalletBookmark.styled';
+import { BookmarkLanding } from '../../components/BookmarkLanding';
 
 export const WalletBookmark = () => {
-  const { t } = useTranslation();
   return (
-    <PageContainer>
-      <Alert severity="info">{t('info.message.fundsToExchange')}</Alert>
-      <Card sx={{ marginTop: '16px' }}>
-        <FormControl
-          fullWidth
-          sx={{
-            paddingTop: '6px',
-            paddingBottom: '5px',
-          }}
-        >
-          <Input
-            sx={{ height: '96px', alignItems: 'flex-start' }}
-            size="small"
-            autoComplete="off"
-            autoCorrect="off"
-            autoCapitalize="off"
-            spellCheck="false"
-            placeholder={t('bookmark.enterAddressOrEns') as string}
-          />
-        </FormControl>
-      </Card>
-    </PageContainer>
+    <BookmarkLanding />
   );
 };
